python_program/Backend/requirements.txt
asgiref==3.4.1
dj-database-url==2.1.0
Django==3.2.22
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==4.8.3
packaging==21.3
psycopg2==2.9.1
pyparsing==3.1.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
